Restructured import statements so firm_ce works as a package
