+++ main.py
@@ -1,6 +1,6 @@
 import time
 
-from firm_ce import Model
+from src.firm_ce.model import Model
 
 start_time = time.time()
 model = Model()
