+++ src/firm_ce/common/exceptions.py
@@ -1,4 +1,4 @@
-from firm_ce.common.jit_overload import njit
+from .jit_overload import njit
 
 
 @njit
