+++ src/firm_ce/common/typing.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Tuple
 import numpy as np
 from numpy.typing import NDArray
 
-from firm_ce.common.constants import JIT_ENABLED
+from .constants import JIT_ENABLED
 
 EvaluationRecord_Type = Tuple[str, str, float, float, float, NDArray[np.float64]]
 BroadOptimumVars_Type = Tuple[int, str, bool, str]
