+++ src/firm_ce/constructors/component_cons.py
@@ -2,9 +2,8 @@ from typing import Dict
 
 import numpy as np
 
-from firm_ce.common.typing import DictType, TypedDict, int64
-from firm_ce.constructors.cost_cons import construct_UnitCost_object
-from firm_ce.system.components import (
+from ..common.typing import DictType, TypedDict, int64
+from ..system.components import (
     Fleet,
     Fleet_InstanceType,
     Fuel,
@@ -14,7 +13,8 @@ from firm_ce.system.components import (
     Storage,
     Storage_InstanceType,
 )
-from firm_ce.system.topology import Line_InstanceType, Node_InstanceType
+from ..system.topology import Line_InstanceType, Node_InstanceType
+from .cost_cons import construct_UnitCost_object
 
 
 def construct_Fuel_object(fuel_dict: Dict[str, str]) -> Fuel_InstanceType:
