+++ src/firm_ce/fast_methods/fleet_m.py
@@ -1,10 +1,10 @@
-from firm_ce.common.constants import FASTMATH
-from firm_ce.common.exceptions import raise_static_modification_error
-from firm_ce.common.jit_overload import njit
-from firm_ce.common.typing import DictType, TypedDict, boolean, float64, int64, unicode_type
-from firm_ce.fast_methods import generator_m, storage_m
-from firm_ce.system.components import Fleet, Fleet_InstanceType, Generator_InstanceType, Storage_InstanceType
-from firm_ce.system.topology import Line_InstanceType, Node_InstanceType
+from ..common.constants import FASTMATH
+from ..common.exceptions import raise_static_modification_error
+from ..common.jit_overload import njit
+from ..common.typing import DictType, TypedDict, boolean, float64, int64, unicode_type
+from ..system.components import Fleet, Fleet_InstanceType, Generator_InstanceType, Storage_InstanceType
+from ..system.topology import Line_InstanceType, Node_InstanceType
+from . import generator_m, storage_m
 
 
 @njit(fastmath=FASTMATH)
