+++ src/firm_ce/optimisation/single_time.py
@@ -2,14 +2,14 @@ import time
 
 import numpy as np
 
-from firm_ce.common.constants import JIT_ENABLED, NUM_THREADS, PENALTY_MULTIPLIER
-from firm_ce.common.jit_overload import jitclass, njit, prange
-from firm_ce.common.typing import boolean, float64, unicode_type
-from firm_ce.fast_methods import fleet_m, generator_m, line_m, network_m, static_m, storage_m
-from firm_ce.optimisation.balancing import balance_for_period
-from firm_ce.system.components import Fleet_InstanceType
-from firm_ce.system.parameters import ScenarioParameters_InstanceType
-from firm_ce.system.topology import Network_InstanceType
+from ..common.constants import JIT_ENABLED, NUM_THREADS, PENALTY_MULTIPLIER
+from ..common.jit_overload import jitclass, njit, prange
+from ..common.typing import boolean, float64, unicode_type
+from ..fast_methods import fleet_m, generator_m, line_m, network_m, static_m, storage_m
+from ..system.components import Fleet_InstanceType
+from ..system.parameters import ScenarioParameters_InstanceType
+from ..system.topology import Network_InstanceType
+from .balancing import balance_for_period
 
 if JIT_ENABLED:
     from numba import set_num_threads
