+++ src/firm_ce/optimisation/solver.py
@@ -8,8 +8,11 @@ import numpy as np
 from numpy.typing import NDArray
 from scipy.optimize import OptimizeResult, differential_evolution
 
-from firm_ce.common.constants import SAVE_POPULATION
-from firm_ce.optimisation.broad_optimum import (
+from ..common.constants import SAVE_POPULATION
+from ..system.components import Fleet_InstanceType, Generator_InstanceType, Storage_InstanceType
+from ..system.parameters import ModelConfig, ScenarioParameters_InstanceType
+from ..system.topology import Line_InstanceType, Network_InstanceType
+from .broad_optimum import (
     append_to_midpoint_csv,
     broad_optimum_objective,
     build_broad_optimum_var_info,
@@ -19,10 +22,7 @@ from firm_ce.optimisation.broad_optimum import (
     write_broad_optimum_bands,
     write_broad_optimum_records,
 )
-from firm_ce.optimisation.single_time import Solution, evaluate_vectorised_xs
-from firm_ce.system.components import Fleet_InstanceType, Generator_InstanceType, Storage_InstanceType
-from firm_ce.system.parameters import ModelConfig, ScenarioParameters_InstanceType
-from firm_ce.system.topology import Line_InstanceType, Network_InstanceType
+from .single_time import Solution, evaluate_vectorised_xs
 
 
 class Solver:
