+++ src/firm_ce/system/components.py
@@ -1,10 +1,10 @@
 import numpy as np
 
-from firm_ce.common.constants import JIT_ENABLED
-from firm_ce.common.jit_overload import jitclass
-from firm_ce.common.typing import DictType, boolean, float64, int64, unicode_type
-from firm_ce.system.costs import LTCosts, LTCosts_InstanceType, UnitCost_InstanceType
-from firm_ce.system.topology import Line_InstanceType, Node_InstanceType
+from ..common.constants import JIT_ENABLED
+from ..common.jit_overload import jitclass
+from ..common.typing import DictType, boolean, float64, int64, unicode_type
+from .costs import LTCosts, LTCosts_InstanceType, UnitCost_InstanceType
+from .topology import Line_InstanceType, Node_InstanceType
 
 if JIT_ENABLED:
     fuel_spec = [
