+++ src/firm_ce/system/costs.py
@@ -1,6 +1,6 @@
-from firm_ce.common.constants import JIT_ENABLED
-from firm_ce.common.jit_overload import jitclass
-from firm_ce.common.typing import float64, int64
+from ..common.constants import JIT_ENABLED
+from ..common.jit_overload import jitclass
+from ..common.typing import float64, int64
 
 if JIT_ENABLED:
     unitcost_spec = [
